# perf: reuse buffer for hashes (root calculation and leaves push)