--- a/hasher.go
+++ b/hasher.go
@@ -58,6 +58,7 @@ type NmtHasher struct { //nolint:revive
 	// "HashNode".
 	ignoreMaxNs      bool
 	precomputedMaxNs namespace.ID
+	buffer           *byteBuffer
 
 	tp   byte   // keeps type of NMT node to be hashed
 	data []byte // written data of the NMT node
@@ -188,10 +189,9 @@ func (n *NmtHasher) HashLeaf(ndata []byte) ([]byte, error) {
 	if err := n.ValidateLeaf(ndata); err != nil {
 		return nil, err
 	}
-
 	nID := ndata[:n.NamespaceLen]
 	resLen := int(2*n.NamespaceLen) + n.baseHasher.Size()
-	minMaxNIDs := make([]byte, 0, resLen)
+	minMaxNIDs := n.getBytes(resLen)
 	minMaxNIDs = append(minMaxNIDs, nID...) // nID
 	minMaxNIDs = append(minMaxNIDs, nID...) // nID || nID
 
@@ -307,8 +307,8 @@ func (n *NmtHasher) HashNode(left, right []byte) ([]byte, error) {
 
 	// compute the namespace range of the parent node
 	minNs, maxNs := computeNsRange(lRange.Min, lRange.Max, rRange.Min, rRange.Max, n.ignoreMaxNs, n.precomputedMaxNs)
-
-	res := make([]byte, 0, len(minNs)+len(maxNs)+h.Size())
+	totalLen := len(minNs) + len(maxNs) + h.Size()
+	res := n.getBytes(totalLen)
 	res = append(res, minNs...)
 	res = append(res, maxNs...)
 
@@ -316,6 +316,23 @@ func (n *NmtHasher) HashNode(left, right []byte) ([]byte, error) {
 	h.Write(left)
 	h.Write(right)
 	return h.Sum(res), nil
+}
+
+func (n *NmtHasher) initBuffer() {
+	n.buffer = newBuffer()
+}
+
+func (n *NmtHasher) resetBuffer() {
+	if n.buffer != nil {
+		n.buffer.reset()
+	}
+}
+
+func (n *NmtHasher) getBytes(size int) []byte {
+	if n.buffer != nil {
+		return n.buffer.get(size)
+	}
+	return make([]byte, 0, size)
 }
 
 // computeNsRange computes the namespace range of the parent node based on the namespace ranges of its left and right children.
@@ -326,4 +343,42 @@ func computeNsRange(leftMinNs, leftMaxNs, rightMinNs, rightMaxNs []byte, ignoreM
 		maxNs = leftMaxNs
 	}
 	return minNs, maxNs
+}
+
+// byteBuffer is a simple non-thread-safe buffer of byte slices
+type byteBuffer struct {
+	free      [][]byte // available buffers
+	allocated [][]byte // buffers currently in use
+}
+
+// newBuffer creates a new byte buffer
+func newBuffer() *byteBuffer {
+	return &byteBuffer{}
+}
+
+// get retrieves a byte slice from the buffer
+// If no free buffers available, allocates a new one
+func (p *byteBuffer) get(size int) []byte {
+	var b []byte
+
+	if len(p.free) > 0 {
+		lastIdx := len(p.free) - 1
+		b = p.free[lastIdx]
+		p.free = p.free[:lastIdx]
+
+		if cap(b) < size {
+			b = make([]byte, size)
+		}
+	} else {
+		b = make([]byte, size)
+	}
+	b = b[:0]
+	p.allocated = append(p.allocated, b)
+	return b
+}
+
+// reset moves all allocated buffers back to free buffer
+func (p *byteBuffer) reset() {
+	p.free = append(p.free, p.allocated...)
+	p.allocated = p.allocated[:0]
 }