--- a/hasher_test.go
+++ b/hasher_test.go
@@ -8,10 +8,9 @@ import (
 	"reflect"
 	"testing"
 
+	"github.com/celestiaorg/nmt/namespace"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-
-	"github.com/celestiaorg/nmt/namespace"
 )
 
 const (