## Overview

**This PR introduces an option `ReuseBuffers`** which will enable the `NamespacedMerkleTree` and the `NmtHasher` to reuse the buffers. The pipeline looks like this (if the option is enabled):
- on first run we use the hasher normally, but underneath it tracks all the slices allocated for hashes
- after we call `Reset` on nmt, the buffers are reset and the previously allocated slices are available for reuse
- the next time the allocations will not happen.

This gives us roughly 15% performance gains when used repeatedly, results in 3x less memory used and 95% less allocations.

**If the option is not enabled, the tree functionality is not affected.**

## Benchmarks

```
cpu: Apple M1 Max
BenchmarkNMTReuseVsNoReuse
BenchmarkNMTReuseVsNoReuse/128-leaves-Root-NoReuse
BenchmarkNMTReuseVsNoReuse/128-leaves-Root-NoReuse-10         	   16126	     73970 ns/op	   54377 B/op	     405 allocs/op
BenchmarkNMTReuseVsNoReuse/128-leaves-Root-ReuseTree
BenchmarkNMTReuseVsNoReuse/128-leaves-Root-ReuseTree-10       	   17980	     67043 ns/op	   19035 B/op	      16 allocs/op
BenchmarkNMTReuseVsNoReuse/256-leaves-Root-NoReuse
BenchmarkNMTReuseVsNoReuse/256-leaves-Root-NoReuse-10         	    8052	    156591 ns/op	  114570 B/op	     793 allocs/op
BenchmarkNMTReuseVsNoReuse/256-leaves-Root-ReuseTree
BenchmarkNMTReuseVsNoReuse/256-leaves-Root-ReuseTree-10       	    9129	    133909 ns/op	   37509 B/op	      18 allocs/op
BenchmarkNMTReuseVsNoReuse/512-leaves-Root-NoReuse
BenchmarkNMTReuseVsNoReuse/512-leaves-Root-NoReuse-10         	    3771	    316411 ns/op	  240045 B/op	    1565 allocs/op
BenchmarkNMTReuseVsNoReuse/512-leaves-Root-ReuseTree
BenchmarkNMTReuseVsNoReuse/512-leaves-Root-ReuseTree-10       	    4594	    264981 ns/op	   78535 B/op	      20 allocs/op
BenchmarkNMTReuseVsNoReuse/1024-leaves-Root-NoReuse
BenchmarkNMTReuseVsNoReuse/1024-leaves-Root-NoReuse-10        	    1806	    660065 ns/op	  545798 B/op	    3110 allocs/op
BenchmarkNMTReuseVsNoReuse/1024-leaves-Root-ReuseTree
BenchmarkNMTReuseVsNoReuse/1024-leaves-Root-ReuseTree-10      	    2260	    540510 ns/op	  160695 B/op	      26 allocs/op
BenchmarkNMTReuseVsNoReuse/2048-leaves-Root-NoReuse
BenchmarkNMTReuseVsNoReuse/2048-leaves-Root-NoReuse-10        	     877	   1376183 ns/op	 1037494 B/op	    6194 allocs/op
BenchmarkNMTReuseVsNoReuse/2048-leaves-Root-ReuseTree
BenchmarkNMTReuseVsNoReuse/2048-leaves-Root-ReuseTree-10      	    1076	   1084137 ns/op	  325373 B/op	      38 allocs/op
```